delete vertex neighbours

+++ src/llist.c
@@ -20,7 +20,7 @@ void llist_delete(llist_t* list) {
     free(list);
 }
 
-int llist_length(llist_t* list) {
+size_t llist_length(llist_t* list) {
     llist_t* buffer = list;
     int size = 0;
     while (buffer != NULL && buffer->value != NULL){
